refactor: create src folder and run test folder with tap
