+++ .gitignore
@@ -26,6 +26,9 @@ coverage
 # nyc test coverage
 .nyc_output
 
+# tap test coverage
+.tap
+
 # Grunt intermediate storage (https://gruntjs.com/creating-plugins#storing-task-files)
 .grunt
