+++ src/routes/example/index.ts
