+++ test/helper.ts
@@ -9,7 +9,7 @@ import { fileURLToPath } from "node:url";
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
-const AppPath = path.join(__dirname, "../app.js");
+const AppPath = path.join(__dirname, "../src/app.ts");
 
 // Fill in this config with all the configurations
 // needed for testing the application
