+++ test/plugins/support.test.ts
@@ -1,8 +1,8 @@
-import { test } from "node:test";
+import { test } from "tap";
 import assert from "node:assert";
 
 import Fastify from "fastify";
-import Support from "../../plugins/support.js";
+import Support from "../../src/plugins/support.js";
 
 test("support works standalone", async () => {
   const fastify = Fastify();
