+++ tsconfig.json
@@ -1,6 +1,8 @@
 {
   "extends": "fastify-tsconfig",
   "compilerOptions": {
-    "outDir": "dist"
-  }
+    "outDir": "dist",
+    "typeRoots": ["@types", "node_modules/@types"]
+  },
+  "include": ["@types", "src/**/*.ts"]
 }
